virtio-blk: add support for zoned block devices
This PoC patch adds support for Zoned Block Devices (ZBDs) to
the kernel virtio-blk driver.

The patch accompanies the virtio-blk ZBD support draft that is now
being proposed for standardization. The draft can be found at
<insert the link to oasis-tcs mailing list archive here>.

In terms of testing, at this point this code is compile-only as the
QEMU part is to be implemented by an intern as a part of Outreachy
community mentorship initiative. This code might undergo some changes
once the QEMU part becomes available.

Since virtio uses a fixed command size for all block requests, it
became necessary to use a larger request structure compared to the one
used for regular block devices to create some room for some flags that
are defined in the draft for zone operations and, most importantly, to
make room for carrying the Zoned Append sector value from the device
back to the driver along with the request status.

The patch introduces dedicated ZBD handlers for request queuing and
completion. This way, it becomes possible to support ZBDs without
adding any extra code to the main data path that is used for "regular"
aka non-zoned devices.

The zone-specific code in the patch is heavily influenced by NVMe ZNS
code in drivers/nvme/host/zns.c, but it is simpler because the proposed
virtio ZBD draft only covers the zoned device features that are
relevant to the zoned functionality provided by Linux block layer.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>